Reflect Node version compatability
Node 13.x and 14.x appear incompatible at present with the google-cloud/firestore transitive dependency, due to an issue with GRPC. I believe there is a work-around, by changing compiler flags and rebuilding etc. For a script such as this, probably easier to use nvm etc to use 12.x for this.